pyproject.toml
[project]
name = "auto-hinge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dearpygui>=2.0.0",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pyautogui>=0.9.54",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
]
